bot/requirements.txt
beautifulsoup4
psycopg2-binary
aiogram